requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
abseil-cpp=20211102.0=he9d5cce_0
absl-py=1.4.0=py311hecd8cb5_0
aiohttp=3.8.3=py311h6c40b1e_0
aiosignal=1.2.0=pyhd3eb1b0_0
async-timeout=4.0.2=py311hecd8cb5_0
attrs=22.1.0=py311hecd8cb5_0
blas=1.0=mkl
blinker=1.4=py311hecd8cb5_0
brotlipy=0.7.0=py311h6c40b1e_1002
bzip2=1.0.8=h1de35cc_0
c-ares=1.19.0=h6c40b1e_0
ca-certificates=2023.05.30=hecd8cb5_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2023.7.22=py311hecd8cb5_0
cffi=1.15.1=py311h6c40b1e_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py311hecd8cb5_0
cryptography=41.0.2=py311h3b477ad_0
ffmpeg=4.2.2=h97e5cf8_0
filelock=3.9.0=py311hecd8cb5_0
freetype=2.12.1=hd8bbffd_0
frozenlist=1.3.3=py311h6c40b1e_0
gettext=0.21.0=he85b6c0_1
giflib=5.2.1=h6c40b1e_3
gmp=6.2.1=he9d5cce_3
gmpy2=2.1.2=py311h1c2e9e1_0
gnutls=3.6.15=hed9c0bf_0
google-auth=2.6.0=pyhd3eb1b0_0
google-auth-oauthlib=0.5.2=py311hecd8cb5_0
grpc-cpp=1.48.2=h4ed1731_1
grpcio=1.48.2=py311h4ed1731_1
icu=58.2=h0a44026_3
idna=3.4=py311hecd8cb5_0
intel-openmp=2023.1.0=ha357a0b_43547
jinja2=3.1.2=py311hecd8cb5_0
jpeg=9e=h6c40b1e_1
lame=3.100=h1de35cc_0
lcms2=2.12=hf1fd2bf_0
lerc=3.0=he9d5cce_0
libcxx=14.0.6=h9765a3e_0
libdeflate=1.17=hb664fd8_0
libffi=3.4.4=hecd8cb5_0
libiconv=1.16=hca72f7f_2
libidn2=2.3.4=h6c40b1e_0
libopus=1.3.1=h1de35cc_0
libpng=1.6.39=h6c40b1e_0
libprotobuf=3.20.3=hfff2838_0
libtasn1=4.19.0=h6c40b1e_0
libtiff=4.5.0=hcec6c5f_2
libunistring=0.9.10=h9ed2024_0
libvpx=1.7.0=h378b8a2_0
libwebp=1.2.4=hf6ce154_1
libwebp-base=1.2.4=h6c40b1e_1
libxml2=2.10.3=h930c0e2_0
llvm-openmp=14.0.6=h0dcd299_0
lz4-c=1.9.4=hcec6c5f_0
markdown=3.4.1=py311hecd8cb5_0
markupsafe=2.1.1=py311h6c40b1e_0
mkl=2023.1.0=h8e150cf_43559
mkl-service=2.4.0=py311h6c40b1e_1
mkl_fft=1.3.6=py311hdb55bb0_1
mkl_random=1.2.2=py311hdb55bb0_1
mpc=1.1.0=h6ef4df4_1
mpfr=4.0.2=h9066e36_1
mpmath=1.3.0=py311hecd8cb5_0
multidict=6.0.2=py311h6c40b1e_0
ncurses=6.4=hcec6c5f_0
nettle=3.7.3=h230ac6f_1
networkx=3.1=py311hecd8cb5_0
numpy=1.25.2=py311h728a8a3_0
numpy-base=1.25.2=py311h53bf9ac_0
oauthlib=3.2.2=py311hecd8cb5_0
openh264=2.1.1=h8346a28_0
openssl=3.0.10=hca72f7f_0
pillow=9.4.0=py311hcec6c5f_0
pip=23.2.1=py311hecd8cb5_0
protobuf=3.20.3=py311hcec6c5f_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pyjwt=2.4.0=py311hecd8cb5_0
pyopenssl=23.2.0=py311hecd8cb5_0
pysocks=1.7.1=py311hecd8cb5_0
python=3.11.3=hf27a42d_1
pytorch=2.0.1=py3.11_0
re2=2022.04.01=he9d5cce_0
readline=8.2=hca72f7f_0
requests=2.31.0=py311hecd8cb5_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
setuptools=68.0.0=py311hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h6c40b1e_0
sympy=1.11.1=py311hecd8cb5_0
tbb=2021.8.0=ha357a0b_0
tensorboard=2.12.1=py311_0
tensorboard-data-server=0.7.0=py311h7242b5c_0
tensorboard-plugin-wit=1.6.0=py_0
tk=8.6.12=h5d9f67b_0
torchaudio=2.0.2=py311_cpu
torchvision=0.15.2=py311_cpu
typing_extensions=4.7.1=py311hecd8cb5_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.16=py311hecd8cb5_0
werkzeug=2.2.3=py311hecd8cb5_0
wheel=0.38.4=py311hecd8cb5_0
x264=1!157.20191217=h1de35cc_0
xz=5.4.2=h6c40b1e_0
yarl=1.8.1=py311h6c40b1e_0
zlib=1.2.13=h4dc903c_0
zstd=1.5.5=hc035e20_0
